requirements.txt
flask
flask_cors
pymongo
google-generativeai
newsapi-python
schedule
python-dotenv
